Added deployment for staging builds

+++ deploy.sh
@@ -3,7 +3,12 @@
 # Bash safety settings
 set -eu -o pipefail
 
-DEPLOY_SSH_PATH=${DEPLOY_USER}@${DEPLOY_HOST}:${DEPLOY_DIRECTORY}
+# For staging builds deploy in /tmp
+if [[ $1 == 'staging' ]]; then
+    DEPLOY_DIRECTORY="/tmp"
+fi
+
+DEPLOY_SSH_PATH="${DEPLOY_USER}@${DEPLOY_HOST}:${DEPLOY_DIRECTORY}"
 
 # Deploy the content of the public folder
 rsync -r --quiet --delete-after ${TRAVIS_BUILD_DIR}/public/* ${DEPLOY_SSH_PATH}/public/
@@ -12,12 +17,15 @@ rsync --quite --delete-after ${TRAVIS_BUILD_DIR}/docker-compose.yml ${DEPLOY_SSH
 # Deploy the .env file created during the CI build
 rsync --quite --delete-after ${TRAVIS_BUILD_DIR}/.env ${DEPLOY_SSH_PATH}
 
+if [[ $1 == 'production' ]]; then
 # Restart the server container
 ssh ${DEPLOY_USER}@${DEPLOY_HOST} <<EOF
 cd ${DEPLOY_DIRECTORY}
 docker-compose down
 EOF
+
 ssh ${DEPLOY_USER}@${DEPLOY_HOST} <<EOF
 cd ${DEPLOY_DIRECTORY}
 docker-compose up -d nginx-server
 EOF
+fi
